Extend set_slice() to allow control of upsert behavior.
Calls to set_slice() are forced to use the upsert model of data
manipulation.  This can be a highly desirable editing model, but it also
has disadvantages, particularly in the context of the 'set' and 'autoset'
form actions, which imply--but don't enforce--a distinction between
insert and update at the data-storage level.

The effects on insert can be particularly insidious, where one faces
having the assumed behavior on duplicate primary keys thwarted by a
surreptitious conversion to a SQL update. In such an instance, instead of
existing data being protected by a duplicate PK error, the extant record
is replaced by the data from the insert. Detecting this condition,
particularly on a large table, is virtually impossible.

While the results of an update-to-insert adjustment are more benign, it
still presents a nuisance if a strict update is intended. To correct it,
one merely needs to delete the newly created, unanticipated row, and such
behavior does not destroy existing data.

To enforce update or insert, set_slice()'s $key arg can be optionally
passed as an array ref, essentially replacing the existing call:

  set_slice($key,$fary,$vary)

with

  set_slice([$opt, $key],$fary,$vary)

where $opt->{dml} is set to the desired value. $opt as a hash ref is
used so that any possible future opt-style params can simply be loaded
into the existing calling structure.

Change details:

* Default behavior for set_slice() is 'upsert'. If you do nothing to
  your code or catalog, the behavior remains unchanged.

* $opt->{dml} can be 'insert', 'update', or anything else. If it's
  anything else, it has no specific behavior currently. It defaults the
  value 'upsert' just to be somewhat self-documenting and open up the
  possibility of behaviors based on that value in the future.

* Despite the decision to key it off of 'dml', this has no effect on
  deletes. Deletes have no ambiguous behavior, reflected in the fact that
  deletes have their own dedicated method.

* Change only has core impact when processing requests through
  Vend::Data::update_data(). However, any direct calls to set_slice() may
  avail themselves of the new feature simply by overloading the $key arg
  in the same fashion.

* Behavior of Vend::Data::update_data() can be in three modes,
  controllable by the new 'dml' pragma.

  + No pragma setting works in current mode, with upsert behavior.

  + Pragma 'dml=preserve' restricts inserts to insert-only, but allows
    the fall-through behavior from update to insert. As the name preserve
    implies, it means no existing data can be clobbered.

    The advantage to preserve is the easy use of set_slice() as a
    record-cloning operation. In the table editor, one can clone a
    record by simply changing the PK. Without this behavior, one must
    completely re-enter existing data to the new key's name to clone.

  + Pragma 'dml=strict' forces update or insert to only perform the
    requested action.