Remove unnecessary calls to item_subtotal and thus the database.
In the first chunk, when there's a discount, there were extra calls to
item_subtotal.

In the second chunk, apply_discount only takes one argument, and it calls
item_subtotal itself, so there's no point in calling it in the parent.

Patch by Josh Braegger <jbraegger@backcountry.com>, with a little cleanup
and removal of the now-unused $tmp variable by myself.

Thanks, Josh!