Fix design flaw in the new child-process tag:
* Need to double-fork and close sockets and filehandles to disconnect from
the parent process so that the parent doesn't wait for the child to finish.

* Add a new routine &Vend::Server::cleanup_for_exec to handle clearing the
private socket variables, usable from other child code as well.

* Can no longer return a useful PID because of the double-fork, so return
nothing from the tag.


Some other tweaks as well:

* Change process description and allow it to be customized, especially
useful for long-running processes.

* Don't bother retrying if fork attempts fail, as that gets prohibitively
complicated on double-forks and probably indicates a suffocating system
we shouldn't add load to anyway.

* For now, at least, don't bother with optional single-fork implementation
which is mostly useful for a tiny bit of improved performance that's not
worth the hassle for a background process.

* Ignore not just an empty body, but also one with only whitespace.

* Remove documentation for unimplemented umask option.