* Allow catalogs to be set to have Perl always global by default.
	AllowGlobal        catname
	PerlAlwaysGlobal   catname

  This is a global directive (i.e. interchange.cfg).

  [perl global=0] will still be honored, i.e. that will be interpreted
  by Safe.

* Allow catalogs to turn off "strict" in global mode by default:

	PerlNoStrict       catname

  This is a global directive (i.e. interchange.cfg).

This is intended as an easy way to allow catalogs to work all right
with Vend::Charset. Sad, but can't think of any better way short of
maintaining our own version of the UTF8 modules. It is really sad,
because the Perl powers that be have totally abandoned Opcode and
Safe.

NOT RECOMMENDED FOR USE BY NON-SOPHISTICATED INTERCHANGE USERS.
But I bet that will be ignored....