Correct issue with hi-bit characters in search strings.
This corrects the "Wide character in subroutine entry" error that occurs when hi-bit
characters are used in a search. The failure was caused by Digest::MD5's reluctance
to process characters > 0xFF, so we just convert any search options to UTF8 before
calculating the MD5.

Courtesy of David Christensen <david@endpoint.com>.