Add encoding/fallback options to Vend::File::readfile/writefile.
These options were specifically designed to be backwards-compatible
with both legacy interchange (i.e., no special encoding handling
specified) and the existing MV_UTF8 handling, while allowing the
ability to explicitly override either encoding or fallback method in
either case.

The fallback method was chosen to be minimally intrusive while still
allowing the developer to notice when there are issues with the
encoding when chosen.