Vend::Payment::PayflowPro - gateway module for PayPal Payflow Pro HTTPS POST
Paypal will be phasing out the use of the old Verisign pfpro interface. This
module offers (almost) a drop-in replacement for Vend::Payment::Signio. There
may be minor adjustments required to the payment route, and the gateway param
must be changed from signio (or verisign) to payflowpro.

Original framework from Tom Tucker.