* Add support for SHA1 encrypted userdb passwords.
* Add new "promote" feature. When active, and passwords
  of any of the other algorithms are present, on next
  login the user's password will be promoted to the
  target hashing algorithm. This way, password strength
  can be increased organically.

Use of SHA1 passwords can be specified in the same manner
as currently MD5 can be:

UserDB  ui  sha1  1

To utilize the promotion feature, you add a similar line
for the UserDB definition:

UserDB  ui  promote 1

Promote implies that strength is increased, but in reality
promotion will move in any direction desired. The requested
hashing algorithm is the target, and whatever the form of the
passwords in the database, they will be converted to the target.

E.g., if neither sha1 nor md5 is specified, and the database
currently has md5 passwords, if promote is added, it will have
the effect of promoting to crypt(), the target hashing algorithm
(which happens to be the default).

If promote is not used, the change is fully backward compatible.
Whatever method is specified will be used, and if the database
has passwords of a different algorithm, authentication will fail.

You should not specify more than 1 hashing type. If you specify
both md5 and sha1, you'll be subject to the whims of hash
ordering from keys().

Also note that, before promoting to a stronger hash, you should
ensure your database's password field is long enough to hold the
new, longer datum.

Original work from Steven Jenkins <steven@endpoint.com> for
framework of promotion code.