New Global Timeout feature for payment gateways.
The different payment gateways all tend to implement their own timeout features
with varying degrees of success. In particular, LWP-based clients do not appear
to have any capability of passing a timeout by the developer whenever it uses
the https protocol. Thus, the use of LWP's timeout feature is illusory and
forces any activity to a 180s timeout.

The new feature completely separates out the gateway activity with a fork,
giving it reliable control over the duration of the gateway request. It also
eliminates the need for each gateway module to implement its own timeout,
either because the developer chose not to, didn't consider it, or because the
developer discovered it was hopeless to do so using LWP.

The feature is invoked using the "global_timeout" option in the payment route,
or as an opt passed to [charge]. global_timeout should be any positive integer,
which will define the number of seconds until the timeout is triggered. If this
new option is not used, the effect is a no-op, with behavior identical to that
prior to the feature. No changes in config means completely backward
compatible.

Additionally, a new "global_timeout_msg" option is available so that the
message produced if the alarm fires can be customized within the payment route.

E.g.:

    Route   payflowpro    id                 "__PAYFLOWPRO_ID__"
    Route   payflowpro    secret             "__PAYFLOWPRO_SECRET__"
    Route   payflowpro    partner            "__PAYFLOWPRO_PARTNER__"
    Route   payflowpro    vendor             "__PAYFLOWPRO_VENDOR__"
    Route   payflowpro    host               "__PAYFLOWPRO_SERVER__"
    Route   payflowpro    transaction        A
    Route   payflowpro    global_timeout     20
    Route   payflowpro    global_timeout_msg "We're sorry, but due to a system error, we are unable to process online orders at this time. Please call __TOLLFREE__ and let one of our customer-service specialists assist you. We apologize for the inconvenience."