New directive SourceCookie - support for persistent affiliate tracking. Initial framework developed by Peter Ajamian <peter@pajamian.dhs.org>.
Setting SourceCookie defines the relevant attributes of a cookie to be
maintained in conjunction with the usual session-only parameter
$Session->{source}. Its usage eliminates the duration of the user's session as
the limiting factor for applying credit to a referal.

SourceCookie and SourcePriority would be expected to typically work in tandem,
and thus the same cookie defined in both contexts. However, there is no such
requirement to do so. Defining SourceCookie by itself merely makes the cookie
available any time the core source routines set or manipulate
$Session->{source}. That cookie may, or may not, be leveraged as an element in
SourcePriority, which itself may look to a cookie not maintained by
SourceCookie.

SourceCookie supports the following attributes:

* name (required)
* expire (any format supported for [set-cookie])
* domain
* path
* secure
* autoreset

autoreset is a boolean that, when true, will cause each request from the client
to reset the cookie in the response, effectively refreshing the expiration time
relative to the current time. Uses, for example, might include a desire to
ensure that the source cookie last "forever" (autoreset + sufficiently long
expire period) or for more obscure uses such as "Affiliate should last [value
of expire] from the last request".

Attributes may be positional in order of (name expire domain path secure), but
it is recommended that they be expressed as key=value pairs for clarity.

Example using both SourcePriority and SourceCookie together in catalog.cfg:

SourcePriority mv_pc mv_source cookie-MV_SOURCE
SourceCookie name=MV_SOURCE expire="180 days"