New user_merge specialsub is run from the [user-merge] usertag.  Arguments passed to the sub are: ($from_user, $from_urec, $to_user, $to_urec, $udb, $tdb) with the following values:
$from_user - The username of the user being merged from.
$from_urec - Hashref of columns in the userdb table for the user being merged
             from.
$to_user   - The username of the user being merged to.
$to_urec   - Hashref of columns in the userdb table for the user being merged
             to.  Any changes made to this hashref will be recorded back into
             the userdb for this user.
$udb       - $db object for the userdb table.
$tdb       - $db object for the transactions table.

The return value should be false to indicate that normal processing of the merge
should continue as normal for this from user (continue on to merge carts and
transactions) or true to indicate that no further processing should be done for
this user (do not merge the carts or transactions).  A false return value will
also cause any changes to $to_urec to be recorded into the userdb for the to
user at the conclusion of the merge.  Note that if there are multiple from users
then this specialsub will be run for each of the from users.

Example:

SpecialSub user_merge user_merge_sub
Sub user_merge_sub <<EOS
sub {
	my ($from_user, $from_urec, $to_user, $to_urec, $udb, $tdb) = @_;

	# Copy first and last name over from from user:
	$to_urec->{fname} = $from_urec->{fname};
	$to_urec->{lname} = $from_urec->{lname};

	# False return value tells the [user-merge] tag to save the new $to_urec
	# structure off as well as to continue with normal processing of the
	# merge.
	return;
}
EOS