Remove CVV2/CSC from default credit card encrypted block template
The card security code should not be stored at all, even in encrypted
form. This makes the default behavior compliant with section 3.2.2 of
PCI-DSS 1.2:

https://www.pcisecuritystandards.org/security_standards/download.html?id=pci_dss_v1-2.pdf

It is of course still possible to manually supply a template that
stores the card security code in violation of PCI-DSS requirements, so
developers should review any custom credit card encryption templates
to make sure that the CVV2 is not included, and purge it from any
historical data they have stored.

Thanks to Mark Lipscombe for calling attention to this.