Allow Interchange daemon to start with no Catalog declarations
When you start Interchange without specifying a SocketFile and also
no Catalog directives in interchange.cfg, the daemon dies with this
seemingly unrelated error:

No sockets -- INTERCHANGE SERVER TERMINATING

That was because the default @$Global::SocketFile array happened to be
populated by the catalog postprocessing routines, which are never called
when no catalog is defined.

We (Stefan, Davor, and I) moved this default setting to
&Vend::Server::run_server instead, which ensures it always runs, and the
Interchange daemon can start without any catalogs.