Abort daemon startup when required module is missing and clean up error output
Fix problem with eval $@ error result's scope in global Perl module
require routine. This was caused because logGlobal contains an eval
itself that overrides $@. Now when a "Require module Something::Special"
directive is issued and not satisfied, it is fatal as was originally
intended.

Remove logGlobal call that results in duplicate error output.

Correctly say "Aborting Interchange daemon" instead of "Aborting
catalog" when dying on global config errors.