Always log route_order errors regardless of errors_to setting
Slightly reorganizes the error handling path in route_order to always
log the collected errors to ::logError even when errors_to is set.  If
errors_to is set and there is an error/failure within send_email, the
errors will essentially be lost since they never made it to their
intended recipients.  Changing the behavior to always log this ensures
that the errors will not go unnoticed.