Fixed bug that causes options to not show in some places.
This removes the checks for the o_enable field in the master options table
record which we no longer write to the table for new options from etc/report,
etc/receipt.html and etc/mail_receipt.  The check has been replaced with a check
for the option_type field in the products table.  Note that option_type is
actually hard-coded in this instance rather than checking
$Config->{OptionsEnable} as the hard coding is much more efficient.  It might
pay to set this field name in a catalog variable, though, instead.