Change log level 'auth' (which doesn't exist) to 'warning'
'auth' is a facility, not a level.

The erroneous 'auth' level setting was formerly a no-op because 'auth'
wouldn't exist in a correctly-configured SysLog hash, but with the new
logging code it actually passed through as intended and raised an error
both in /usr/bin/logger and Sys::Syslog calls.