Add Vend::CharSet::mime_name for installations which lack proper Encode.pm support
Not all versions of Encode in IC-supported perl versions support the
mime_name method.  This patch adds checking for the existance of the
method, and provides a simple replacement in the case that it lacks
it.

The main goal here is to support earlier versions of Encode.pm, while
still allowing us to normalize and use the utf8 charset.

Additional special-cases can be added to Vend::CharSet::mime_name as needed.