Fix css.tag to properly output the css when using the inline <style> block
css.tag attempts to write a file out to the filesystem after reading
in the css via either variable or literal.  If the file path it
attempts to write to is not writable, for whatever reason, instead of
creating a <link> tag to the written file, it attempts to create a
<style> tag containing the css.

Currently, if it ever creates the style tag, it will never contain the
css.  When the location is not writable, it skips the portion of code
that reads in the actual css, either from the literal option or the
contents of the variable.

This patch moves the reading of the css up to a point where it can't
be skipped, allowing both the link and style tags to be created
properly.

Report and patch by Justin Otten <justin.lasotten@gmail.com>