Add checks for MV_UTF8 in global variable space
Depending on the context in which a particular piece of code is
called, $::Variable will be aliased to either the variables defined in
interchange.cfg (i.e., $Global::Variable) or in the specific
catalog.cfg.

In order to handle the fact that MV_UTF8 can be defined in either
interchange.cfg or catalog.cfg, change all checks to look at both
$::Variable and $Global::Variable explicitly.

This is known to have affected the [import] tag at the very least, and
likely had other subtle implications in other places.