Add MSIE and Gecko UA strings to the default list of NotRobotUA patterns
This will correct a string of false positives found when checking the
list of RobotUAs against the wide variety of toolbars found in the
wild in User-Agent strings.

This does open up the possibility of more obscure bots getting
assigned sessions from the appserver, but this was deemed to be a
reasonable tradeoff given the current bot/toolbar ecosystem.