Trust only the last hop before a trusted proxy, when multiples exist
The original implementation here trusted that the X-Forwarded-For header
was trustworthy, and when multiple intermediate proxies (or forgeries)
were present, you could end up with an internal IP address on the remote
network listed, or something entirely bogus.

When multiple proxies are present, trust only the one that connected to
our trusted proxy. Even though that doesn't get us to the "original" IP
address, it's reliable and public, and the one we would've seen if we
weren't using TrustProxy at all.