Add new pragma url_no_session_id to suppress session IDs in generated URLs
For any site that already requires cookies to function, you can add this
pragma to the catalog:

    Pragma url_no_session_id

To disable in a particular page or template file:

    [tag pragma no_image_rewrite 0][/tag]

It ensures that Interchange-generated URLs won't include mv_session_id,
id, or mv_pc URL query string parameters.

This is mostly of interest for search engine optimization (SEO) purposes,
but is also helpful for making more pages cacheable by a proxy, etc.