Allow country_select widget use with Autofill programs
The country_select widget creates a javascript function depending on
the country selected, in order to display the proper "state set" for
the chosen country using a select widget with the name
"state_cs_in".

A problem arises however when auto form fillers are used because
instead of populating the value of the hidden "state" input field as
you would expect, they populate the value of the created select
widget named "state_cs_in".  To the user, it appears that the proper
state has been selected, however at this point only the
"state_cs_in" has been set. Because a javascript function to set the
"state" value to the chosen "state_cs_in" selected value is never
triggered with an "onChange" when the "auto-fill" is used, the value
never gets set, and the user submits the form and receives a

State/Province /('' not a two-letter state code)/

error message.

A simple solution for this problem is to add code to the
mv_form_profile that performs the work if "state_cs_in" has been
changed but "state" has not.