Fix for processing GDBM files while using UTF8
It appears that utf8 filter should not be applied when creating GDBM
files. This seems to cause a double encoding. Prerequisite is that
the TXT file out of which the GDBM file is built is UTF8 encoded.

But when working with UTF8 this should be the case else the TXT files
are not able to be worked with with various languages.