Enable case-insensitivity in UserDB for indirect_login.
This patch allows catalogs that are using the indirect_login feature to
combine that with ignore_case to enable case-insensitive logins.

A common use-case is to have email address be the indirect login field, so
one thing to be aware of is that it's legal for two separate e-mail
addresses to differ in capitalization only (e.g. user@domain is distinct
from User@domain).