Only apply GDBM filters if they are not already installed
In certain circumstances, using the same GDBM-based file in multiple catalogs with
GDBM_ENABLE_UTF8=1 set was causing the GDBM filters to be installed multiple times on the same
handle, which was resulting in encoding/decoding errors.

Vend::Table::GDBM will now only install the filters if they do not exist on this handle, which
should preserve existing behavior plus account for these corner-cases.