Force scalar context for is_yes and is_no
This was exposed by an obscure bug in the Vend::UserDB constructor when $opt->{cgi} was defined but
did not match the underlying regex; since it was constructed in list context this resulted in an
odd-number of hash elements.

I reviewed all uses of is_yes and is_no in the code, none of which depend on being called in array
context.

Thanks to Jeff Boes <jeff@endpoint.com> for reporting.