Increase SEO value of more links.
New body tags for [more-list] can be used to "salt" more links
with SEO-driven content. These tags can turn a link like:

http://foo.com/scan/MM=[md5:x:x:x].html?mv_more_ip=1&mv_nextpage=results&pf=sql

into

http://foo.com/scan/leisure-bikes/traditional-bikes/page-2/MM=[md5:x:x:x].html?mv_more_ip=1&mv_nextpage=results&pf=sql

While the new tags will work inside any more list, they are really
only valuable in conjunction with the use of Permanent More, so
those higher-value links cached by search engines actually lead
to viable resources. See release notes and documentation on
Permanent More if you are unfamiliar with it, or you can see my blog
post on it at:

http://blog.endpoint.com/2012/01/interchange-search-caching-with.html

The new body tags are:

* [more-pretty-url]

This container tag includes the data of interest you'd like to plug
into the more-list URLs to make them "pretty". The value can contain
anything matching a-z, 0-9, /, _, and -. If any characters other
than those described are included, they are converted in contiguous
blocks to a single -. Any contiguous multiples of - (after
conversion) and / are collapsed into a single one. Finally, any -
and / are stripped from the head and tail.

What this does is allow a flexible vat into which you can throw
meaningful data (without having to test or sanitize at all) which
will produce a path fragment of relevant terms that is put into the
more-list URL between the "scan/" and "/MM=..." paths within overall
URL.

Example:

[more-list]
[more-pretty-url]men's clothing/pants/bermuda shorts[/more-pretty-url]
...
[/more-list]

Produces a more link with the following format:

http://foo.com/scan/men-s-clothing/pants/bermuda-shorts/MM=a52a53...html?...

Even though the body of a [more-list] is interpolated on reparse,
the contents within [more-pretty-url] specifically are interpolated.
It is expected that you would have some algorithm you would use to
determine the pretty parameters and those would be rendered via ITL
within the [more-pretty-url] tag.

* [more-incl-pageno]

Additionally, to provide more context and uniqueness to the URLs,
you can optionally request that the page number be included as the
last path segment of the "pretty" URL. This can either be the
default (which is "page-N"), or can be any customized string you
define, provided that you include the sprintf %d integer
interpolation flag so that the page number will be included.

Unlike [more-pretty-url], the contents of this tag are *not*
pre-interpolated.

Expanding the above example:

[more-list]
[more-pretty-url]men's clothing/pants/bermuda shorts[/more-pretty-url]
[more-incl-pageno][/more-incl-pageno]
...
[/more-list]

Produces a more link with the following format for page 4:

http://foo.com/scan/men-s-clothing/pants/bermuda-shorts/page-4/MM=a52a53...html?...

Or a customized page path with:

[more-list]
[more-pretty-url]men's clothing/pants/bermuda shorts[/more-pretty-url]
[more-incl-pageno]search page %d[/more-incl-pageno]
...
[/more-list]

produces new URL for page 4:

http://foo.com/scan/men-s-clothing/pants/bermuda-shorts/search-page-4/MM=a52a53...html?...