Patch [area] to accept decoded strings.
is_utf8 is badly name. First, it's an internal function, but it's
used all over the place in IC.

From the doc:

    is_utf8(STRING [, CHECK])
      [INTERNAL] Tests whether the UTF8 flag is turned on in the STRING. If
      CHECK is true, also checks the data in STRING for being well-formed
      UTF-8. Returns true if successful, false otherwise.

So, basically, is_utf8 returns true if the string is decoded, not when
is encoded in utf8.

use strict;
use warnings;
use utf8;
ok(is_utf8("à"));
diag "When we encode, the string is not \"utf8\" any more";
ok(!is_utf8(encode(utf8 => 'à')));