Allow DirectoryIndex to find index page in root pages/ dir:
It seems DirectoryIndex, when the readin() sub doesn't find a page
for '', adds a slash, so it then looks for '/index'. This is wrong
when acting on the root pages/ directory, so now we check if the
page is blank, and if so skip the slash.

Came to light when attempting to serve a root index.html page via
PageTables.