Revert "Embed Safe 2.07 into Vend::Safe to avoid various problems with recent versions of Safe."
This is broken in at least perl 5.20, possibly earlier; I also don't believe that this is a good
approach to take, particularly as the Safe module relies on specific internal perl modules which we
are also currently not including (nor could we effectively).

I *would* be interested to look at the issues that this commit was intended to fix, to see if we can
come up with a better general-purpose solution which works across multiple versions of perl and
Safe.

This reverts commit 6264540e5d33313d412c3c1899d1452b3e7ac311.