Fix error creating indexes in MySQL and other dbs when QUOTE_IDENTIFIERS is set
The problem was that Vend::Table::DBI was assembling an index name based on the
quoted table name plus '_' plus quoted column name, resulting in a syntactically
invalid name like `variants`_`sku`, leading to errors.  Instead assemble the
name unquoted, and quote the final result.

This is related to commit 86fbcb14c402efda89b2329b73270ae513715ddf which fixes
the same issue for PostgreSQL, the error also needed correcting for the other db
types as well.