Avoid quoting $column in set_field().
Avoid modifying $column by quoting it in the middle of set_field().  This had
the unintended consequence that later changes to the code which tried to access
column settings were broken due to the need to use $column as a key.  We now
create $qcolumn which holds the quoted value instead.