Do not substitute key for non-existing columns when QUOTE_IDENTIFIERS is set.
There is a bug in get_slice and set_slice where the key column (or more
specifically the column at index 0) is substituted when QUOTE_IDENTIFIERS is
set.  This fixes the bug to push through the quoted version of whatever is
passed instead (and will likely error out from the db transaction instead).
Erroring out here is much more appropriate than attempting to read or write
seemingly random garbage to/from the db.