Fixed a bug in the [area] tag when a full URL is passed as the href.
When a fully-qualified URL is passed in the href attribute of the [area] tag and
the form attribute is also used Interchange would overwrite the
protocol://domain portion of the URL with the one in the VendURL configuration
directive.

Please note that if your code relied on this buggy behaviour then you will need
to fix it before upgrading.