Correct logic of DisplayErrors
Most directives (and other places with this specific directive) use 'or' logic to allow you to
define directives in either `interchange.cfg` or `catalog.cfg` to enable a default-off behavior at
either the Interchange server level or at the level of a specific catalog.

As originally coded, the DisplayErrors directive would have only worked when *both*
`interchange.cfg` and `catalog.cfg` had this enabled.

Change this so it shows up when either the current catalog or the global config has this set.

Since the old default was 'no', someone intended this information to be output, so this is the
correct behavior, considering that I only fixed the output of this message in the previous commit.